:recycle: useEvaluateExpression 로 리팩토링
