+++ lib/views/game/magi-mixer/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./useMathField";
+export * from "./useEvaluateExpression";
